refactor: move drag-handle into framework packages as sub-imports
Remove `@tiptap/extension-drag-handle-react` and `@tiptap/extension-drag-handle-vue-3`
packages and consolidate their functionality as sub-imports within the main framework
packages, following the same pattern as menu components.

Changes:
- Create `src/drag-handle` directories in `@tiptap/react` and `@tiptap/vue`
- Move DragHandle components from separate packages into new directories
- Add `./drag-handle` exports to both framework packages
- Update all demo imports to use new sub-import paths:
  - `@tiptap/react/drag-handle` (React)
  - `@tiptap/vue/drag-handle` (Vue)
- Update tsup configs to build drag-handle entry points
- Add `@tiptap/extension-drag-handle` as optional dependency
- Remove deprecated packages

BREAKING CHANGE: DragHandle is now imported from framework packages instead of
separate extension packages. Update imports accordingly.
