--- packages/extension-drag-handle-vue-3/README.md
@@ -1,18 +1,0 @@
-# @tiptap/extension-drag-handle-vue3
-
-[![Version](https://img.shields.io/npm/v/@tiptap/extension-drag-handle-vue3.svg?label=version)](https://www.npmjs.com/package/@tiptap/extension-drag-handle-vue3)
-[![Downloads](https://img.shields.io/npm/dm/@tiptap/extension-drag-handle-vue3.svg)](https://npmcharts.com/compare/tiptap?minimal=true)
-[![License](https://img.shields.io/npm/l/@tiptap/extension-drag-handle-vue3.svg)](https://www.npmjs.com/package/@tiptap/extension-drag-handle-vue3)
-[![Sponsor](https://img.shields.io/static/v1?label=Sponsor&message=%E2%9D%A4&logo=GitHub)](https://github.com/sponsors/ueberdosis)
-
-## Introduction
-
-Tiptap is a headless wrapper around [ProseMirror](https://ProseMirror.net) – a toolkit for building rich text WYSIWYG editors, which is already in use at many well-known companies such as _New York Times_, _The Guardian_ or _Atlassian_.
-
-## Official Documentation
-
-Documentation can be found on the [Tiptap website](https://tiptap.dev).
-
-## License
-
-Tiptap is open sourced software licensed under the [MIT license](https://github.com/ueberdosis/tiptap/blob/main/LICENSE.md).
