--- packages/extension-drag-handle-vue-3/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@tiptap/extension-drag-handle-vue-3",
-  "description": "drag handle extension for tiptap with vue 3",
-  "version": "3.20.0",
-  "homepage": "https://tiptap.dev",
-  "keywords": [
-    "tiptap",
-    "tiptap extension"
-  ],
-  "license": "MIT",
-  "funding": {
-    "type": "github",
-    "url": "https://github.com/sponsors/ueberdosis"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ueberdosis/tiptap",
-    "directory": "packages/extension-drag-handle-vue-3"
-  },
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": {
-        "import": "./dist/index.d.ts",
-        "require": "./dist/index.d.cts"
-      },
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "src",
-    "dist"
-  ],
-  "peerDependencies": {
-    "@tiptap/extension-drag-handle": "workspace:^",
-    "@tiptap/pm": "workspace:^",
-    "@tiptap/vue": "workspace:^",
-    "vue": "^3.0.0"
-  },
-  "devDependencies": {
-    "@tiptap/extension-drag-handle": "workspace:^",
-    "@tiptap/pm": "workspace:^",
-    "@tiptap/vue": "workspace:^",
-    "vue": "^3.0.0"
-  },
-  "scripts": {
-    "build": "tsup",
-    "lint": "prettier ./src/ --check && eslint --cache --quiet --no-error-on-unmatched-pattern ./src/"
-  }
-}
