--- packages/extension-drag-handle-vue-3/src/index.ts
@@ -1,5 +1,0 @@
-import { DragHandle } from './DragHandle.js'
-
-export * from './DragHandle.js'
-
-export default DragHandle
