--- packages/extension-drag-handle-vue-3/tsup.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'tsup'
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  tsconfig: '../../tsconfig.build.json',
-  outDir: 'dist',
-  dts: true,
-  clean: true,
-  sourcemap: true,
-  format: ['esm', 'cjs'],
-})
