+++ packages/react/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig(
-  ['src/index.ts', 'src/menus/index.ts'].map(entry => ({
+  ['src/index.ts', 'src/menus/index.ts', 'src/drag-handle/index.ts'].map(entry => ({
     entry: [entry],
     tsconfig: '../../tsconfig.build.json',
     outDir: `dist${entry.replace('src', '').split('/').slice(0, -1).join('/')}`,
