+++ packages/vue/src/drag-handle/index.ts
@@ -1,0 +1,1 @@
+export * from './DragHandle.js'
